+++ src/main/java/com/imsweb/datagenerator/hl7/rule/ObservationRequestSegmentRule.java
@@ -9,6 +9,7 @@ import java.util.Map;
 
 import com.imsweb.datagenerator.hl7.NaaccrHl7DataGeneratorOptions;
 import com.imsweb.datagenerator.hl7.NaaccrHl7DataGeneratorRule;
+import com.imsweb.datagenerator.utils.DistributionUtils;
 import com.imsweb.datagenerator.utils.RandomUtils;
 import com.imsweb.layout.hl7.Hl7MessageBuilder;
 import com.imsweb.layout.hl7.entity.Hl7Message;
@@ -33,16 +34,16 @@ public class ObservationRequestSegmentRule extends NaaccrHl7DataGeneratorRule {
                 .withField(4, "11529-5", "SURGICAL PATH REPORT", "LN")
 
                 // OBR-10: collector identifier
-                .withField(10, "COLLECTOR-ID", "COLLECTOR-LAST", "COLLECTOR-FIRST")
+                .withField(10, RandomUtils.getRandomStringOfDigits(5), DistributionUtils.getNameLast(), DistributionUtils.getNameFirst())
 
                 // OBR-14: specimen received date/time
                 .withField(14, DateTimeFormatter.ofPattern("yyyyMMdd").format(RandomUtils.getRandomDateBetween(LocalDate.now().minusYears(1L), LocalDate.now())))
 
                 // OBR-16: ordering provider
                 .withField(16)
-                .withComponent(2, "PHYS-LAST")
-                .withComponent(3, "PHYS-FIRST")
-                .withComponent(4, "M")
+                .withComponent(2, DistributionUtils.getNameLast())
+                .withComponent(3, DistributionUtils.getNameFirst())
+                .withComponent(4, RandomUtils.getRandomStringOfLetters(1))
 
                 // OBR-17: order call back phone
                 .withField(17)
@@ -51,6 +52,6 @@ public class ObservationRequestSegmentRule extends NaaccrHl7DataGeneratorRule {
 
                 // OBR-32: principal result interpreter
                 .withField(32)
-                .withComponent(1, null, "PHYS-LAST", "PHYS-FIRST", "M");
+                .withComponent(1, null, DistributionUtils.getNameLast(), DistributionUtils.getNameFirst(), RandomUtils.getRandomStringOfLetters(1));
     }
 }