+++ CHANGELOG.md
@@ -1,15 +1,19 @@
 ## Synthetic data generator
 
-**Changes in version 0.4**
+**Changes in version 1.0**
 
+ - Added support for generating NAACCR HL7 messages.
  - Added new parameter for Registry ID.
  - Fixed bug where DX date and/or DOLC could be set to a future date.
+ - Updated Layout library from version 1.4 to version 1.5.1.
+ - Updated Algorithm library from version 1.4.4 to version 1.7.
+ - Updated Staging library from version 2.3 to version 2.5.
 
 **Changes in version 0.3**
 
  - Removed Class-Path attribute from generated JAR.
- - Updated layout library from version 1.3.3 to version 1.4.
- - Updated algorithm library from version 1.4 to version 1.4.4.
+ - Updated Layout library from version 1.3.3 to version 1.4.
+ - Updated Algorithm library from version 1.4 to version 1.4.4.
  - Updated Staging library from version 2.0 to version 2.3.
  - Updated Commons IO library from version 2.4 to version 2.5.
  - This library now requires Java 8 at minimum.